Time storage changed from float to double on host.

+++ include/quickgl/InputManager.hpp
@@ -58,7 +58,7 @@ namespace qgl {
 		
 	private:
 		
-		float time, deltaTime;
+		double time, deltaTime;
 	};
 }
